Feature EPL-1731: Fix layout
