+++ packages/MainUI/src/App.tsx
@@ -2,13 +2,16 @@ import { CssBaseline, ThemeProvider } from '@mui/material';
 import { theme } from '@workspaceui/componentlibrary/src/theme';
 import { Outlet } from 'react-router-dom';
 import MetadataProvider from './contexts/metadata';
+import RecordProvider from './contexts/record';
 
 export default function App() {
   return (
     <ThemeProvider theme={theme}>
       <CssBaseline />
       <MetadataProvider>
-        <Outlet />
+        <RecordProvider>
+          <Outlet />
+        </RecordProvider>
       </MetadataProvider>
     </ThemeProvider>
   );
