+++ packages/MainUI/src/hooks/useRecordContext.ts
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import { RecordContext } from '../contexts/record';
+
+const useRecordContext = () => useContext(RecordContext);
+
+export default useRecordContext;
